Move field-related stuff to separate file
