+++ dora/src/field.rs
@@ -1,0 +1,48 @@
+use std::ops::{Index, IndexMut};
+
+use crate::ty::BuiltinType;
+use dora_parser::interner::Name;
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub struct FieldId(usize);
+
+impl FieldId {
+    pub fn idx(self) -> usize {
+        self.0
+    }
+}
+
+impl From<usize> for FieldId {
+    fn from(data: usize) -> FieldId {
+        FieldId(data)
+    }
+}
+
+#[derive(Debug)]
+pub struct Field {
+    pub id: FieldId,
+    pub name: Name,
+    pub ty: BuiltinType,
+    pub offset: i32,
+    pub reassignable: bool,
+}
+
+impl Index<FieldId> for Vec<Field> {
+    type Output = Field;
+
+    fn index(&self, index: FieldId) -> &Field {
+        &self[index.0]
+    }
+}
+
+impl IndexMut<FieldId> for Vec<Field> {
+    fn index_mut(&mut self, index: FieldId) -> &mut Field {
+        &mut self[index.0]
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct FieldDef {
+    pub offset: i32,
+    pub ty: BuiltinType,
+}
