+++ dora/src/lib.rs
@@ -27,6 +27,7 @@ mod dseg;
 mod error;
 mod exception;
 mod execstate;
+mod field;
 mod gc;
 mod handle;
 mod masm;
