+++ dora/src/vm.rs
@@ -15,10 +15,11 @@ use crate::baseline::dora_native::{self, InternalFct, InternalFctDescriptor, Nat
 use crate::baseline::dora_throw;
 use crate::baseline::fct::{JitFct, JitFctId};
 use crate::baseline::map::{CodeDescriptor, CodeMap};
-use crate::class::{Class, ClassDef, ClassDefId, ClassId, FieldId};
+use crate::class::{Class, ClassDef, ClassDefId, ClassId};
 use crate::driver::cmd::Args;
 use crate::error::diag::Diagnostic;
 use crate::exception::DoraToNativeInfo;
+use crate::field::FieldId;
 use crate::gc::{Address, Gc};
 use crate::object::{Ref, Testing};
 use crate::safepoint::{PollingPage, Safepoint};
